+++ Library/lib/controls/types.ts
@@ -1260,7 +1260,7 @@ export type DesktopDockSettings = {
   autoHide: boolean;
   position: "bottom" | "left" | "right";
   resizable: boolean;
-  size: number;
+  size: "small" | "medium" | "large";
 };
 
 export type DesktopEnvironmentSettings = {